Correct changes in numeric types between Python 2 and 3.
* Change syntax used to express octal numbers
* Replace calls to long() with int().  Python is smart enough to convert
  to long transparently

+++ src/subscription_manager/gui/mysubstab.py
@@ -141,7 +141,7 @@ class MySubscriptionsTab(widgets.SubscriptionManagerTab):
         if not response:
             return
 
-        serial = long(selection['serial'])
+        serial = int(selection['serial'])
 
         if self.identity.is_valid():
             self.pb = progress.Progress(_("Removing"),
@@ -269,7 +269,7 @@ class MySubscriptionsTab(widgets.SubscriptionManagerTab):
         if not serial:
             return
 
-        cert = self.entitlement_dir.find(long(serial))
+        cert = self.entitlement_dir.find(int(serial))
         if not cert:
             return
